requirements.txt
streamlit==1.31.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.1
ultralytics==8.0.196
pyyaml>=5.3.1
tqdm>=4.64.0
shutil
